# Fix code generation in the presence of constants and zeros