@lassepe 2023-09-03
The CI failure seems unrelated. At least I cannot reproduce it locally.