@brianguenter 2023-09-08
Looked at your latest changes. The code is cleaner, very nice, but there appears to be a substantial performance regression for the case of an all constant matrix with mixed zeros and non-zeros. Maybe for other cases as well, haven't checked all the combinations.

Used this function to create the constant matrices:
```julia
function setup(n)
    FD = FastDifferentiation
    @variables y
    expr = rand(n, n)
    exprzc = map(x -> x > 0.5 ? FD.Node(x) : FD.Node(0.0), expr)
    return make_function(exprzc, FD.Node[]), make_function(exprzc, FD.Node[], in_place=true)
end
```
<details> 
  <summary> Version info </summary>

```julia
Julia Version 1.9.3
Commit bed2cd540a (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 32 × AMD Ryzen 9 7950X 16-Core Processor
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, znver3)
  Threads: 1 on 32 virtual cores
Environment:
  JULIA_EDITOR = code.cmd
  JULIA_NUM_THREADS = 1
```
</details>

These are the timings for the PR code

```julia
#PR version
julia> @benchmark bin(res,Float64[])
BenchmarkTools.Trial: 9259 samples with 1 evaluation.
 Range (min … max):  516.000 μs …  1.176 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     531.800 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   537.253 μs ± 19.081 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

         ▄█▃▂
  ▁▁▁▁▁▄▇████▆▅▄▃▃▃▃▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  516 μs          Histogram: frequency by time          612 μs <

 Memory estimate: 96 bytes, allocs estimate: 3.

julia> @benchmark bout(Float64[])
BenchmarkTools.Trial: 2994 samples with 1 evaluation.
 Range (min … max):  1.317 ms …   7.239 ms  ┊ GC (min … max):  0.00% … 55.26%
 Time  (median):     1.369 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   1.669 ms ± 681.910 μs  ┊ GC (mean ± σ):  12.32% ± 16.82%

  █▇▄▅▃▁                     ▃▅▂▁▁                            ▁
  ██████▇▆▁▃▄▁▁▁▁▃▁▃▁▁▁▁▅▁▁▁▃██████▆▅▁▁▄▁▁▃▁▁▃▁▁▅▇▇█▆▇▆▆▅▅▁▄▅ █
  1.32 ms      Histogram: log(frequency) by time      3.81 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 5.
```
These are the timings for the old code
```julia
#Old code
julia> bout,bin = setup(1000);

julia> res = rand(1000,1000);

julia> @benchmark bin(res,Float64[])
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  269.500 μs … 362.000 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     270.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   270.757 μs ±   3.231 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

   █▇▄▂                                                         ▁
  █████▇▆▆▆████▆▇▇▆▆▇▆▅▆▅▅▆▆▆▅▅▆▅▆▆▅▆▆▅▆▄▅▄▆▄▅▄▄▃▅▃▃▃▄▅▅▂▃▅▃▅▄▅ █
  270 μs        Histogram: log(frequency) by time        286 μs <

 Memory estimate: 96 bytes, allocs estimate: 3.

julia> @benchmark bout(Float64[])
BenchmarkTools.Trial: 5930 samples with 1 evaluation.
 Range (min … max):  659.100 μs …   2.242 ms  ┊ GC (min … max):  0.00% … 42.06%
 Time  (median):     688.200 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   841.760 μs ± 278.571 μs  ┊ GC (mean ± σ):  13.85% ± 18.39%

  ▆█▅▃▃▂▁▁▁    ▁▃▃▂                     ▄▃▃▁                    ▁
  ██████████████████▆▇▆▆▄▅▄▆▆▄▆▄▆▁▄▆▅▁▇███████▆▇▇▆▆▄▄▆▆▅▆▆▇▆▆▆▆ █
  659 μs        Histogram: log(frequency) by time       1.73 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 3.
```
Looked at the generated code to see where the slowdown might be coming from. There are several things that might be contributing. The `result` array is being traversed twice, once for the constants and once for the zeros. And these traversals are less regular than the old code since they skip elements. The old code just copies an array of constants.

```julia
#PR code
(RuntimeGeneratedFunctions.RuntimeGeneratedFunction{(:input_variables,), FastDifferentiation.var"#_RGF_ModTag", FastDifferentiation.var"#_RGF_ModTag", (0x13987dd6, 0xbe179260, 0x154e1a7c, 0xb363840f, 0x8e14b797), Expr}(quote
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:180 =#
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:180 =# @inbounds begin
            #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:181 =#
            begin
                result_element_type = promote_type(Float64, eltype(input_variables))
                result = Array{result_element_type}(undef, (4, 4))
                result[CartesianIndex{2}[CartesianIndex(1, 1), CartesianIndex(2, 1), CartesianIndex(3, 1), CartesianIndex(4, 1), CartesianIndex(1, 2), CartesianIndex(4, 2), CartesianIndex(2, 3), CartesianIndex(4, 3), CartesianIndex(1, 4), CartesianIndex(2, 4), CartesianIndex(3, 4)]] .= zero(result_element_type)
                result[CartesianIndex{2}[CartesianIndex(2, 2), CartesianIndex(3, 2), CartesianIndex(1, 3), CartesianIndex(3, 3), CartesianIndex(4, 4)]] .= [0.7954135366237669, 0.5530606683573553, 0.5151457247633523, 0.533598585953031, 0.6427988879183532]
                return result
            end
        end
end), 

RuntimeGeneratedFunctions.RuntimeGeneratedFunction{(:result, :input_variables), FastDifferentiation.var"#_RGF_ModTag", FastDifferentiation.var"#_RGF_ModTag", (0x251bfd6e, 0x70196b2a, 0x6c72e358, 0x35b108db, 0xc65fe188), Expr}(quote
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:176 =#
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:176 =# @inbounds begin
            #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:177 =#
            begin
                result_element_type = eltype(input_variables)
                result[CartesianIndex{2}[CartesianIndex(1, 1), CartesianIndex(2, 1), CartesianIndex(3, 1), CartesianIndex(4, 1), CartesianIndex(1, 2), CartesianIndex(4, 2), CartesianIndex(2, 3), CartesianIndex(4, 3), CartesianIndex(1, 4), CartesianIndex(2, 4), CartesianIndex(3, 4)]] .= zero(result_element_type)
                result[CartesianIndex{2}[CartesianIndex(2, 2), CartesianIndex(3, 2), CartesianIndex(1, 3), CartesianIndex(3, 3), CartesianIndex(4, 4)]] .= [0.7954135366237669, 0.5530606683573553, 0.5151457247633523, 0.533598585953031, 0.6427988879183532]
                return nothing
            end
        end
end))
```

Old code:
```julia
#Old Code
julia> bout,bin = setup(4)
(RuntimeGeneratedFunctions.RuntimeGeneratedFunction{(:input_variables,), FastDifferentiation.var"#_RGF_ModTag", FastDifferentiation.var"#_RGF_ModTag", (0x9fa9db12, 0x332bedc5, 0x180cd068, 0xee1b0483, 0x31b19011), Expr}(quote
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:204 =#
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:204 =# @inbounds begin
            #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:205 =#
            begin
                result = copy([0.0 0.8135299148943799 0.0 0.6812918543356454; 0.5583812839340533 0.923236908240091 0.7047065762753687 0.0; 0.0 0.0 0.7731522140458759 0.5189997523232557; 0.6268794344428308 0.0 0.5449987019060242 0.7421889317315941])
                return result
            end
        end
end), 

RuntimeGeneratedFunctions.RuntimeGeneratedFunction{(:result, :input_variables), FastDifferentiation.var"#_RGF_ModTag", FastDifferentiation.var"#_RGF_ModTag", (0x5c5f5c15, 0xcad2eedf, 0xf7a950d5, 0x968ed370, 0x8c9b4746), Expr}(quote
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:200 =#
    #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:200 =# @inbounds begin
            #= c:\Users\seatt\source\FastDifferentiation.jl\src\CodeGeneration.jl:201 =#
            begin
                result .= [0.0 0.8135299148943799 0.0 0.6812918543356454; 0.5583812839340533 0.923236908240091 0.7047065762753687 0.0; 0.0 0.0 0.7731522140458759 0.5189997523232557; 0.6268794344428308 0.0 0.5449987019060242 0.7421889317315941]
            end
        end
end))